Made this change yesterday but just overwrote by pushing WorkingModel4 to mavlabCourse2024
Made this change yesterday but accidentally overwrote it because I just pushed WorkingModel4 branch over to this mavlabCourse2024 branch.

We used the WorkingModel4 branch in our competition. But just now I committed the WorkingModel4 branch to mavlabCourse2024 because the professors will probably look at mavlabCourse2024 branch.

Doing this push caused the readme to be reset back to the original.